pyproject.toml
[tool.poetry]
name = "rawal-backend"
version = "0.1.0"
description = "HR software"
authors = ["Nellow <prabeshpathak002@gmail.com>"]
readme = "README.md"
packages = [{ include = "app" }]
package-mode = false

[tool.poetry.dependencies]
python = ">=3.11,<4.0"
fastapi = ">=0.115.12,<0.116.0"
uvicorn = ">=0.34.0,<0.35.0"
pydantic = ">=2.11.3,<3.0.0"
python-dotenv = ">=1.1.0,<2.0.0"
sqlalchemy = ">=2.0.40,<3.0.0"
psycopg2-binary = ">=2.9.10,<3.0.0"
punq = ">=0.7.0,<0.8.0"
pytest = "^8.3.5"
httpx = "^0.28.1"
pydantic-settings = "^2.8.1"
passlib = {extras = ["bcrypt"], version = "^1.7.4"}
python-jose = {extras = ["cryptography"], version = "^3.4.0"}
alembic = "^1.15.2"
openai = "^1.72.0"
pandas = "^2.2.3"
reportlab = "^4.3.1"
pypdf2 = "^3.0.1"
openpyxl = "^3.1.5"
python-multipart = "^0.0.20"
pycryptodome = "^3.22.0"
tiktoken = "^0.9.0"

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"